requirements.txt
 langchain
openai
requests
flask
python-dotenv
flask-limiter

